feat(test): 重整network目录测试文件结构
将5个测试文件重组为3个标准命名的测试文件：
- 删除孤立文件: final_coverage_test.go, focused_coverage_test.go
- 合并到: interface_test.go (网络接口相关测试)
- 保持89.1%测试覆盖率和完整网络功能测试
- 涵盖HTTP真实IP检测、网络接口管理、IP验证功能
